Add a class that handles direct access to vector elements
Currently we use unions whenever we want to access vector elements
without going through expensive extract() machinery. extract() is not
usable when we want to access more than one element as a general storing
to memory + loading may be more efficient. Also, if the consumer is
within the same functional processor's unit, extract() adds huge
unnecessary overhead which is not possible to optimize out.

This commit adds a class template mem_block, which represents an aligned
memory block with convenient operator[] interface to access its
elements. Data transfer between mem_block and a vector is done using
standard std::memcpy, which is much easier for the compiler to optimize
correctly.