shuffle: Change the semantics of broadcast()
Most 256-bit functions within the library operate as if corresponding
operation was applied to both 128-byte lanes of a vector separately.
broadcast() is inconsistent with this regard: in case of 256-byte
vectors it broadcasts single value to all elements. Because of that,
broadcast() can't be used in generic code that can work on both 128 and
256-bit vectors.

This commit changes the semantics of broadcast() so that in 256-bit
vectors it operates as if two separate broadcasts were applied to the
128-bit elements. Since vector-wide broadcast might be needed in some
cases, a separate broadcast_all() function is added.