shuffle: Rename broadcast_all() to broadcast_w()
It is a good idea to reserve a suffix for consistent naming of functions
that operate on entire vectors (when they are wider than 128 bits)
instead of 128-bit lanes separately. *_all suffix is not appropriate for
that, as it may lead to confusing function names. Thus, we select
another another suffix (*_w, w = wide) and rename the only (recently
created) function it applies to.