memory: Add alignment annotations
In NEON, the load/save intrinsics does not have a way to express the
alignment of arguments. This results in the compiler assuming the worst
case and using instructions for unaligned access, which are roughly two
times slower than normal instructions. This commit adds annotations that
inform the compiler about our interface restrictions. Currently only GCC
(>4.7.0) is supported.