blend: Do not accept generic int128/int256 masks
Primary reason for this commit is preparation for the eventual removal
of int128/int256. Using masks with different element widths is no longer
supported because given the constraints on the mask values it's unlikely
that the mask source will ever have element width that's different from
the element width of the blended values.

This is API breaking change.