removed return statements from void functions
Several of the load_lane methods were declared void, but are returning a value.  I removed the return statements as it appears the api is designed to change the parameter passed by reference.

The float64x2 version of load_lane was not changed as it is declared with a float64x2 return value.