Make scalar arguments work on older compilers
Some compilers such as GCC 4.8 fail to perform template argument
deduction correctly. They fail to match function parameters to function
arguments even if they do not contain types to be deduced. In such case
implicit conversions should be evaluated after the types are deduced,
which is not the case.

Thus for scalar arguments we can't use a wrapper class that may be
implicitly constructed from several types. Unfortunately that means that
for each fundamental type that we want to support a separate overload
must be provided.