Convert scalar args to vector element type (except in bitwise functions)
The first implementation always transferred the bit pattern of the
source scalar argument to all elements of the destination vector. This
was to make cases such as bit_and(v, 0x80000000) work for all vector
types. However, the conversion rules were very unintuitive in the rest
of the cases.

This commit makes the scalar conversions to convert the source value to
the element type of the target vector in the usual C/C++ rules, except
for bitwise functions, where we want to preserve existing behavior.