Store subexpressions as const references in expression objects
Some compilers, such as GCC, can't optimize out all possible cases of
copy construction on some platforms, such as ARM/NEON. This leads to
abysmal performance of expression templates. Storing sub-expressions in
expression templates works around this issue. The caveats are the
following:

* We can never construct new values when building an expression
template. Thus all casts must be done as cast from one reference type to
another (in our case this is not UB, at least in C++11, as the types in
question are standard layout types).

* The user should never store the expression template using auto. Only
conversions to concrete vector type are allowed.