Add bit_select() function
This is similar to blend() but does not require that mask is composed
entirely of either ones or zeros for each element. Currently if this
condition is not satisfied code like the following must be written:

bit_or(bit_and(a, c), bit_andnot(b, c)

Many architectures have a single instruction that performs the same.
Unfortunately it is quite hard for compilers to notice the missing
optimization opportunity. bit_select() helps to avoid that.